added lighweight token authentication to popup messages mutation

+++ app/api/promotions/route.ts
@@ -1,6 +1,8 @@
 import { NextResponse } from 'next/server';
 import Promotion from '@/models/Promotion';
 import connectToDb from '@/lib/mongodb';
+import jwt from 'jsonwebtoken';
+import { getTokenFrom } from '@/app/utils/helpers';
 
 // get the promo data
 export async function GET() {
@@ -17,13 +19,20 @@ export async function GET() {
 // create a new promo (rewrite existing, only 1 allowed)
 export async function POST(request: Request) {
   const body = await request.json();
-  const { onPromotion, text } = body;
 
+  const { onPromotion, text } = body;
   if (onPromotion === undefined || text === undefined) {
     return NextResponse.json({ error: 'missing credentials' }, { status: 400 });
   }
 
+  const token = getTokenFrom(request);
+  if (!token) {
+    return NextResponse.json({ error: 'missing token' }, { status: 401 });
+  }
+
   try {
+    jwt.verify(token, process.env.SECRET!);
+
     await connectToDb();
 
     await Promotion.deleteMany({});
@@ -32,6 +41,7 @@ export async function POST(request: Request) {
 
     return NextResponse.json(result, { status: 201 });
   } catch (error) {
-    console.log(error);
+    console.error(error);
+    return NextResponse.json({ error }, { status: 500 });
   }
 }
